requirements.txt
crewai
crewai-tools
python-dotenv
streamlit
snowflake-connector-python